requirements.txt
youtube-transcript-api==0.6.1
google-api-python-client==2.108.0
anthropic==0.39.0
pandas==2.1.4
python-dotenv==1.0.0
requests
isodate
anthropic